Initial Commit
This is version 1 of my library catalog system. It uses mostly PHP (with
a touch of Javascript) and SQLite3 for the backend. Entries are stored
in 2 tables - books and genres - where books can have multiple genres.
The table can be sorted by the first five columns, or filtered by
read/unread or by genre. Books can be added with the button, or
edited/deleted by clicking the row in the table.